Refactored and handle new storage
